added intake to oi

+++ nohorobo2019code/src/main/java/frc/robot/subsystems/Intake.java
@@ -30,7 +30,7 @@ public class Intake extends Subsystem {
     // Set the default command for a subsystem here.
     // setDefaultCommand(new MySpecialCommand());
   }
-  public void SetRollerMotors(int speed){
+  public void SetRollerMotors(double speed){
     RollerMotor.set(speed);
   }
   public void ExtendRollerSolenoid(){
